price amount becomes a double

+++ src/main/java/gym/reporting/Turnover.java
@@ -4,14 +4,14 @@ import gym.subscriptions.domain.Subscription;
 
 import java.util.List;
 
-public final record Turnover(int total) {
+public final record Turnover(double total) {
 
     public static Turnover monthly(List<Subscription> subscriptions) {
         return new Turnover(
             subscriptions
                 .stream()
                 .map(Subscription::monthlyTurnover)
-                .reduce(0, Integer::sum)
+                .reduce(0.0, Double::sum)
         );
     }
 }
